Refactored the Card container
